backend/requirements.txt
Django==3.2
mixer==7.1.2
Pillow==9.3.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
requests==2.26.0
six==1.16.0
sorl-thumbnail==12.7.0
Faker==12.0.1
djoser==2.1.0
drf-base64==2.0
psycopg2-binary==2.8.6
python-dotenv==0.20.0
django-filter==22.1
gunicorn==20.0.4
